% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MargLikeNAM.R
\name{BFbanam}
\alias{BFbanam}
\title{Bayes factors for Bayesian hypothesis testing under the NAM}
\usage{
BFbanam(
  x,
  hypothesis = NULL,
  prior = NULL,
  complement = TRUE,
  priortype = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{banam} object.}

\item{hypothesis}{A character string containing the informative hypotheses to
evaluate. The default is NULL, which will result in standard exploratory testing
under the model \code{x}.}

\item{prior}{A vector specifying the prior probabilities of the hypotheses.
The default is NULL which will specify equal prior probabilities.}

\item{complement}{a logical specifying whether the complement should be added
to the tested hypothesis under \code{hypothesis}.}

\item{...}{Parameters passed to and from other functions.}
}
\value{
The output is an object of class \code{BF}. The object has elements:
BFtu_exploratory, PHP_exploratory, BFtu_confirmatory, PHP_confirmatory,
BFmatrix_confirmatory, BFtable_confirmatory, PHP_interaction, prior,
hypotheses, estimates, model, call.
}
\description{
The \code{BFbanam} is a function ...
}
\references{
Dittrich, Leender, Mulder, ...
}
