\docType{data}
\name{W_birds}
\alias{W_birds}
\title{Weight matrix for threatened bird data}
\format{A data.frame with 113 rows amd 113 columns}

\description{
Generated data from a study on spatial effect on country's threatened bird rates by McPherson and Nieswiadomy (2005). This data set stores the data on the weighted shared border length of 113 neighboring countries in the year 2000.
}
\usage{
data(W_birds)
}

\references{
McPherson, M. A., & Nieswiadomy, M. L. (2005). Environmental Kuznets curve: threatened species and spatial effects. Ecological Economics, 55(3), 395-407.
}
\keyword{datasets}
