\docType{data}
\name{W_votes}
\alias{W_votes}
\title{Weight matrix for counties in Alabama, US}
\format{
A matrix with 67 rows and 67 columns
}

\description{
This dataset specifies the four nearest neighboring 
counties of 67 counties in Alabama, US. All counties that satisfy 
this criteria were assigned a weight of 0.25.
}
\usage{data(W_votes)}

\references{
Pace, R. K. and R. Barry. 1997. Quick Computation of Spatial Autoregressive Estimators. Geographical Analysis, 29, 232-47. 
Data can be downloaded from 
http://www.spatial-econometrics.com/html/jplv7.zip
}
\keyword{datasets}