\docType{data}
\name{y_birds}
\alias{y_birds}
\title{
Threatened bird rates
}
\format{A vector of the length 113}

\description{
This generated vector contains the threatened bird rates of 113 countries based
on a study by McPherson and Nieswiadomy (2005).
}
\usage{data(y_birds)}

\references{
McPherson, M. A., & Nieswiadomy, M. L. (2005). Environmental Kuznets curve: threatened species and spatial effects. Ecological Economics, 55(3), 395-407.
}
\keyword{datasets}