\docType{data}
\name{y_votes}
\alias{y_votes}
\title{
Logarithmized voter turnout in Alabama, US
}
\format{A vector of the length 67}

\description{
This vector contains the logarithmized voter turnout of 67 counties in Alabama, US.
}
\usage{data(y_votes)}

\references{
Pace, R. K. and R. Barry. 1997. Quick Computation of Spatial Autoregressive
Estimators. Geographical Analysis, 29, 232-47. 
Data can be downloaded from 
http://www.spatial-econometrics.com/html/jplv7.zip
}
\keyword{datasets}